+++ pkg/controllers/componentInstallManager_test.go
@@ -111,7 +111,7 @@ func Test_componentInstallManager_Install(t *testing.T) {
 		// given
 		mockComponentClient := NewMockComponentClient(t)
 		mockComponentClient.EXPECT().UpdateStatusInstalling(context.TODO(), component).Return(component, nil)
-		mockComponentClient.EXPECT().UpdateStatusInstalled(context.TODO(), component).Return(nil, assert.AnError)
+		mockComponentClient.EXPECT().UpdateStatusInstalled(context.TODO(), component).Return(component, assert.AnError)
 		mockComponentClient.EXPECT().AddFinalizer(context.TODO(), component, "component-finalizer").Return(component, nil)
 
 		mockHelmClient := NewMockHelmClient(t)
@@ -128,6 +128,6 @@ func Test_componentInstallManager_Install(t *testing.T) {
 		// then
 		require.Error(t, err)
 		assert.ErrorIs(t, err, assert.AnError)
-		assert.ErrorContains(t, err, "failed to set status installed")
+		assert.ErrorContains(t, err, "failed to update status-installed for component dogu-op")
 	})
 }